Finalização de todas funcionalidades

+++ src/main/java/br/com/fiap/reciclamais/usecase/data/output/PontuacaoBusinessOutput.java
@@ -16,6 +16,6 @@ public class PontuacaoBusinessOutput {
 
     private Double reciclagem;
     private Integer total;
-    private Float percentual;
+    private Double percentual;
     private List<HistoricoBusinessOutput> historico;
 }
